Add a Delete action in "n newest unowned tickets"
Adding a Delete action is necessary for efficient handling of spam tickets.
Clicking "Delete" displays the deleted ticket, at which point the user can
change the status of the ticket if he/she realizes that the deletion was a
mistake.

(Technically, HTTP GETs should not have such side-effects, and these actions
should be HTTP POSTs instead.)